requirements.txt
gradio
google
opencv-python
numpy
tensorflow
mediapipe
python-dotenv
